+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,6 +1,7 @@
 ﻿### Version 3.3.35
 
 - Update: `CSharpConstructor` now supports `IHasCSharpStatements`.
+- Update: CSharpClassMethod can now have an expression body to output the form: `public DateTime GetTime() => DateTime.Now;` 
 
 ### Version 3.3.34